Rename uiApplyPatch to just applyPatch
