+++ src/core/state/RdfState.types.ts
@@ -1,7 +1,7 @@
 import { RdfPatch2 } from '../types';
 import { Store, Quad } from 'n3';
 import { SelectionCallback } from '../types/graphModel';
-import { UiNodeSymbol } from '../ui/uiApplyPatch';
+import { UiNodeSymbol } from '../ui/applyPatch';
 
 export type RdfAction =
 	| { type: 'add'; data: Iterable<Quad> }
