+++ src/core/types/graphModel.ts
@@ -1,5 +1,5 @@
 import { Quad } from 'n3';
-import { Point, UiNodeSymbol } from '../ui/uiApplyPatch';
+import { Point, UiNodeSymbol } from '../ui/applyPatch';
 
 type ElementId = string;
 export type GraphElementBase = {
